Add styling to course, courses datarow and admintools

+++ api/src/main/java/com/example/servermaintenance/account/AccountService.java
@@ -57,7 +57,7 @@ public class AccountService implements UserDetailsService {
         return accounts;
     }
 
-    public Account getAccountById(Long accountId) {
-        return accountRepository.getById(accountId);
+    public Account getAccountById(int accountId) {
+        return accountRepository.getById((long)accountId);
     }
 }
